genai-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
chromadb==0.5.3
numpy==1.26.4
openai==1.25.0
httpx==0.27.0
sentence-transformers==2.7.0
confluent-kafka==2.3.0
python-dotenv==1.2.2
prometheus-fastapi-instrumentator==6.1.0
transformers==4.40.0